feat(reservas): implementar validación de solapamiento y disponibilidad dinámica
- **IReservaRepository**: agregar método `getReservasPorCanchaYFecha`
- **ReservaRepository**:
    - Implementar lógica para evitar reservas que se solapen en la misma cancha y horario
    - Normalizar fechas y generar `slotId` determinista basado en cancha y horario
    - Actualizar mapeo de Firestore para soportar `duracionHoras` como `Double`
- **CanchaViewModel**:
    - Cargar reservas existentes del día al seleccionar cancha o fecha
    - Soportar duraciones fraccionadas (ej. 1.5h) en el flujo de creación
- **CanchaDetailScreen**:
    - **HorariosGrid**: deshabilitar slots que ya están reservados o que ya pasaron (para el día actual)
    - **DuracionSelector**: deshabilitar opciones de duración que causen conflicto con la siguiente reserva existente
- **Modelo**: cambiar el tipo de `duracionHoras` de `Int` a `Double` en `Reserva` y datos de prueba

+++ app/src/main/java/com/example/futbol_tnt/data/model/Reserva.kt
@@ -7,7 +7,7 @@ data class Reserva(
     val usuarioId: String = "",
     val cancha: Cancha,
     val fecha: LocalDateTime,
-    val duracionHoras: Int,
+    val duracionHoras: Double,
     val precioTotal: Double,
     val estado: EstadoReserva,
     val nombreEquipo: String? = null
